back/requirements.txt
openai
langchain~=0.3.3
streamlit~=1.39.0
numpy~=1.26.4
pandas~=2.2.3
langchain-core~=0.3.12
langchain-openai~=0.2.2
pydantic~=2.9.2
fastapi~=0.115.2
uvicorn~=0.32.0
langserve~=0.3.0
requests~=2.32.3
pydantic_core~=2.23.4
python-dotenv